+++ expyriment/stimuli/_rectangle.py
@@ -136,9 +136,9 @@ class Rectangle(Visual):
         return surface
 
     def is_point_inside(self, point_xy):
-        """"DEPRECATED METHOD: Please use 'overlapping_with_position'."""
+        """"OBSOLETE METHOD: Please use 'overlapping_with_position'."""
 
-        return self.overlapping_with_position(point_xy)
+        raise RuntimeError("is_point_inside is an obsolete method. Please use overlapping_with_position")
 
 if __name__ == "__main__":
     from expyriment import control